requirements.txt
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
h11==0.14.0
idna==3.7
outcome==1.3.0.post0
packaging==24.0
PySocks==1.7.1
python-dotenv==1.0.1
python-http-client==3.3.7
PyYAML==6.0.1
requests==2.31.0
schedule==1.2.1
selenium==4.20.0
sendgrid==6.11.0
slack_sdk==3.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
starkbank-ecdsa==2.2.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
webdriver-manager==4.0.1
wsproto==1.2.0
